+++ lakehouse-traces/internal/internalselect/handler_test.go
@@ -80,7 +80,7 @@ func vlMetadataURL(path, version string) string {
 }
 
 func TestHandler_Query_EmptyResult(t *testing.T) {
-	h := NewHandler(&mockStorage{}, 30*time.Second)
+	h := NewHandler(&mockStorage{}, 30*time.Second, nil)
 	mux := http.NewServeMux()
 	h.Register(mux)
 
@@ -105,7 +105,7 @@ func TestHandler_Query_WithDataBlocks(t *testing.T) {
 		},
 	}
 
-	h := NewHandler(store, 30*time.Second)
+	h := NewHandler(store, 30*time.Second, nil)
 	mux := http.NewServeMux()
 	h.Register(mux)
 
@@ -171,7 +171,7 @@ func TestHandler_FieldNames(t *testing.T) {
 		},
 	}
 
-	h := NewHandler(store, 30*time.Second)
+	h := NewHandler(store, 30*time.Second, nil)
 	mux := http.NewServeMux()
 	h.Register(mux)
 
@@ -208,7 +208,7 @@ func TestHandler_FieldValues(t *testing.T) {
 		},
 	}
 
-	h := NewHandler(store, 30*time.Second)
+	h := NewHandler(store, 30*time.Second, nil)
 	mux := http.NewServeMux()
 	h.Register(mux)
 
@@ -233,7 +233,7 @@ func TestHandler_FieldValues(t *testing.T) {
 }
 
 func TestHandler_TenantIDs(t *testing.T) {
-	h := NewHandler(&mockStorage{}, 30*time.Second)
+	h := NewHandler(&mockStorage{}, 30*time.Second, nil)
 	mux := http.NewServeMux()
 	h.Register(mux)
 
@@ -247,7 +247,7 @@ func TestHandler_TenantIDs(t *testing.T) {
 }
 
 func TestHandler_DeleteNoop(t *testing.T) {
-	h := NewHandler(&mockStorage{}, 30*time.Second)
+	h := NewHandler(&mockStorage{}, 30*time.Second, nil)
 	mux := http.NewServeMux()
 	h.Register(mux)
 
@@ -284,7 +284,7 @@ func TestHandler_StreamEndpoints(t *testing.T) {
 		},
 	}
 
-	h := NewHandler(store, 30*time.Second)
+	h := NewHandler(store, 30*time.Second, nil)
 	mux := http.NewServeMux()
 	h.Register(mux)
 
@@ -322,7 +322,7 @@ func TestHandler_StreamEndpoints(t *testing.T) {
 }
 
 func TestHandler_AllEndpointsRegistered(t *testing.T) {
-	h := NewHandler(&mockStorage{}, 30*time.Second)
+	h := NewHandler(&mockStorage{}, 30*time.Second, nil)
 	mux := http.NewServeMux()
 	h.Register(mux)